ar(feat) update users types
