+++ lib/types/currencies.d.ts
@@ -204,3 +204,5 @@ export enum ECryptoCurrencies {
   YFI = 'yearn.finance',
   ZEC = 'Zcash',
 }
+
+export type ECurrency = ETraditionalCurrencies | ECryptoCurrencies;
