+++ lib/types/index.ts
@@ -4,7 +4,10 @@
 export type { IAuthContext, IRMContext, INCharacter, IDCharacter, History } from './contexts';
 
 // model
-export type { UserSchema, User, UserDecoration, OrgDecoration } from './model';
+export type { OrgDecoration } from './model';
+
+// users
+export type { UserSchema, User, UserDecoration } from './users';
 
 // actions
 export type {
@@ -32,7 +35,7 @@ export type {
 export type { IAbility, EAbilityStatus, EUserOrgRoles } from './authorization';
 
 // services
-export type { IFeature } from './services';
+export type { IFeature, IProject, IServiceUserAmbiRelation } from './services';
 
 // system
 export type { Tposition, TpositionX, TpositionY, Tsize, Tthemes } from './atoms';
